[Allomancy] Moves steel and iron lines thread to its own file.
